refactor(edge-client): move the challenge handshake out of edge-identity
`edge-identity.ts` was the `EdgeIdentity` contract plus one response helper;
this branch grew it into something that also does network I/O, which is the
wrong home for it — the file is about *who* is authenticating, not the wire
format they authenticate with.

Split the handshake into `auth-challenge.ts`: header parsing, the `/auth` round
trip, and signing. `edge-identity.ts` is back to just the interface, and the
test file moves with the code.

Public API is unchanged — the barrel re-exports the new module, so external
importers (`plugin-payments` takes `handleAuthChallenge` from the package root)
are untouched. Verified with a plugin-payments build.

No behaviour change; 54 tests unchanged and passing.

(cherry picked from commit 8a91c95b181a35d29d274707830f8444eb64c787)
