+++ packages/core/mesh/edge-client/src/index.ts
@@ -5,6 +5,7 @@
 export * from '@dxos/protocols/buf/dxos/edge/messenger_pb';
 
 export * from './auth';
+export * from './auth-challenge';
 export * from './cors-proxy';
 export * from './defs';
 export * from './edge-client';
